Improve git history command
The default history command, will now only print the tail instead of
the complete history. Moreover, the user can now specify how many
commits of the history, they want to get.

The main reason for this is that the history got too long and telegram
blocked the messages of my bot, and almost always are the last commit
messages the most relevant.