Add, usage for unauthorized users
With these changes, everyone can use this bot without it leaking
sensitive data. Before this change the bot only allowed the admin (me)
access.

While most commands are still exclusive for the admin, as they may leak
data, some are open for public access. This is true for the notification
feature which will be quite handy for some of my classmates.